VS2019 specific fixes, __VA_ARGS__, __attribute__, suppress warnings
* __VA_ARGS__ macro support
* __attribute__ only with __GCC__
* suppress C4101 uninitialized recount warnings, and strcpy insecure warnings.
* no proper const var support
  (with the even weirder error: cannot allocate an array of constant
  size)
* strict dllexport/dllimport decls: fix various omissions

clang-cl can handle it, VS2019 errors. See the new VS2019 CI

+++ src/dwg_api.c
@@ -22861,7 +22861,7 @@ dwg_add_class (Dwg_Data *restrict dwg, const char *const restrict dxfname,
 
 #  define API_UNADD_ENTITY                                                    \
     if (dwg->header.version < R_13b1)                                         \
-      obj->type = -(obj->type);                                               \
+      obj->type = -(long)(obj->type);                                         \
     else                                                                      \
       obj->type = DWG_TYPE_UNUSED
